requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
PyMuPDF==1.23.8
faiss-cpu==1.7.4
numpy==1.26.4
langchain==0.3.27
langchain-experimental==0.3.4
langchain-openai==0.3.29
langchain-community==0.3.27
langchain-core==0.3.74
langchain-text-splitters==0.3.9
beautifulsoup4==4.12.2
python-docx==0.8.11
extract-msg==0.41.1
openai==1.99.6
lxml==4.9.3
pandas==2.3.1
openpyxl==3.1.5
xlrd==2.0.2
rank-bm25==0.2.2
langsmith==0.4.13

